+++ scripts/format.sh
@@ -19,15 +19,11 @@ install_py_packages $py
 python_sources=$(LANG=C find scripts src/bindings/python -type f ! -name "*.md" ! -name "PKG-INFO" -exec file {} \; | grep -F "Python script" | sed 's/:.*//')
 
 # Python formatter
-echo "${python_sources}"
-${py} -m black \
-  --line-length 105 \
-  --skip-string-normalization \
-  --extend-exclude "/(\..*|dist|wheelhouse|.*build|__pycache__|.*\.toml|.*\.egg-info)/" \
-  ${python_sources}
+echo ${python_sources}
+${py} -m ruff check ${python_sources}
 
 # Python linter
-${py} -m flake8 --config setup.cfg ${python_sources}
+${py} -m ruff format ${python_sources}
 
 # clang-format
 echo "Using $(${py} scripts/clang_format_wrapper.py --version)"